pyproject.toml
[project]
name = "meddicc-fse"
version = "0.1.0"
description = ""
authors = [
    {name = "Joseph Matthew Caballas",email = "jmcaballas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=25.1,<26.0)",
]
package-mode = false

[tool.pytest.ini_options]
python_files = [
    "tests.py",
    "test_*.py",
]
DJANGO_SETTINGS_MODULE = "config.settings.test"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"